Add a synthesizer resource
This resource is used at the plug-in's init to start a separate thread for rendering the audio. Here it is used via an Arc as a shared resource, and a Mutex to allow concurrent access to this resource (or at least, to let the compiler allow it while preventing actual concurrent access).
I don't know if this will be performant enough. Let's see.