Reduce size of MIDI-limited values
This will limit the values at compile time. If I make a mistake with the values, the compiler will warn me about it more often then (e.g. negative values or too high pitch). And perhaps it'd be a bit more efficient too, but probably it'll not really have an effect on modern computers.